Ints rather than Bytes. They say it's faster and not bigger.

+++ Greedy/src/main/scala/com/gjos/scala/swoc/Direction.scala
@@ -1,17 +1,17 @@
 package com.gjos.scala.swoc
 
 trait Direction {
-  val x: Byte
-  val y: Byte
+  val x: Int
+  val y: Int
 }
 
 object Direction {
-  case object North extends Direction { val x: Byte = 0; val y: Byte = -1 }
-  case object South extends Direction { val x: Byte = 0; val y: Byte = 1 }
-  case object NorthEast extends Direction { val x: Byte = 1; val y: Byte = 0 }
-  case object SouthEast extends Direction { val x: Byte = 1; val y: Byte = 1 }
-  case object NorthWest extends Direction { val x: Byte = -1; val y: Byte = -1 }
-  case object SouthWest extends Direction { val x: Byte = -1; val y: Byte = 0 }
+  case object North extends Direction { val x: Int = 0; val y: Int = -1 }
+  case object South extends Direction { val x: Int = 0; val y: Int = 1 }
+  case object NorthEast extends Direction { val x: Int = 1; val y: Int = 0 }
+  case object SouthEast extends Direction { val x: Int = 1; val y: Int = 1 }
+  case object NorthWest extends Direction { val x: Int = -1; val y: Int = -1 }
+  case object SouthWest extends Direction { val x: Int = -1; val y: Int = 0 }
 
   def allDirections = List(North, NorthEast, SouthEast, South, SouthWest, NorthWest)
 }
