+++ Greedy/src/main/scala/com/gjos/scala/swoc/protocol/Move.scala
@@ -1,5 +1,5 @@
 package com.gjos.scala.swoc.protocol
 
-case class Move(moveType: MoveType, from: Option[(Byte, Byte)], to: Option[(Byte, Byte)]) {
+case class Move(moveType: MoveType, from: Option[(Int, Int)], to: Option[(Int, Int)]) {
   override def toString = from.map(BoardLocation.label).mkString + s"--$moveType-->" + to.map(BoardLocation.label).mkString
 }
