+++ Greedy/src/main/scala/com/gjos/scala/swoc/protocol/MoveType.scala
@@ -1,21 +1,21 @@
 package com.gjos.scala.swoc.protocol
 
 trait MoveType {
-  val value: Byte
+  val value: Int
 }
 
 object MoveType {
   case object Pass extends MoveType {
-    val value: Byte = 0
+    val value: Int = 0
   }
   case object Attack extends MoveType {
-    val value: Byte = 1
+    val value: Int = 1
   }
   case object Strengthen extends MoveType {
-    val value: Byte = 2
+    val value: Int = 2
   }
 
-  def byValue(i: Byte): MoveType = i match {
+  def byValue(i: Int): MoveType = i match {
     case 0 => Pass
     case 1 => Attack
     case 2 => Strengthen
